Refactor game state classes. GameState -> StandardGameState. StandradGameState -> MoveApplier.
