+++ Application/Program.cs
@@ -1,7 +1,4 @@
 ﻿using GameLogic;
 
-// Test object, created to be sure that the GameLogic project is linked successfully.
-GameState state = new GameState();
-
-HelloWorldPrinter printer = new HelloWorldPrinter();
+var printer = new HelloWorldPrinter();
 printer.Print();
